move duplicated code to the abstract class
